backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
PyJWT==2.10.1
bcrypt==4.3.0
requests==2.32.3
python-dotenv==1.1.0
